pyproject.toml
[project]
name = "ljcourses"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "flask>=3.1.2",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "werkzeug>=3.1.5",
]
